requirements.txt
flask==2.3.3
flask-login==0.6.2
flask-wtf==1.1.1
werkzeug==2.3.7
wtforms==3.0.1
python-dotenv==1.0.0
email-validator==2.0.0